+++ public/index.html
@@ -27,7 +27,7 @@
     <script defer src="/__/firebase/init.js?useEmulator=true"></script>
 
     <!-- <script defer src="server.js" type="module"></script> -->
-    <script defer type="module" src="dist/bundle.js"></script>
+    <script defer src="dist/bundle.js"></script>
   </head>
   <body>
     <div class="header">
@@ -56,8 +56,8 @@
           <div id="stopVideo" class="options__button">
             <i class="fa fa-video-camera"></i>
           </div>
-          <div id="muteButton" class="options__button background__red">
-            <i class="fas fa-microphone-slash"></i>
+          <div id="muteButton" class="options__button">
+            <i class="fas fa-microphone"></i>
           </div>
           <div id="showChat" class="options__button">
             <i class="fa fa-comment"></i>
