+++ webpack.config.js
@@ -12,6 +12,7 @@ module.exports = {
     // map the built code back to the original source format when debugging.
     devtool: "eval-source-map",
     experiments: {
-        topLevelAwait: true,
+        topLevelAwait: false,
     },
+    watch: true,
 }
